src/requirements.txt
pandas
websocket-client
minio